merge

+++ client/src/pages/SellerDashboard/ItemManagement.jsx
@@ -344,8 +344,13 @@ const ItemManagement = () => {
                                 />
                                 <button
                                   className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded"
-                                  onClick={() => {
-                                    handleDeleteItem(item._id);
+                                  onClick={() => { 
+                                    if (
+                                      window.confirm(
+                                        "Confirm Item Deletion?"
+                                      )
+                                    )
+                                      handleDeleteItem(item._id);
                                   }}
                                 >
                                   Delete
